+++ lib/ECommerce/Controllers/AddProductController.php
@@ -1,8 +1,10 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Controllers;
 
+use ECommerce\Site;
+use ECommerce\Tables\Products;
 
 class AddProductController extends Controller {
