+++ lib/ECommerce/Controllers/LoginController.php
@@ -1,8 +1,13 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Controllers;
 
+
+use ECommerce\Site;
+use ECommerce\Tables\Cookies;
+use ECommerce\Tables\Users;
+use ECommerce\User;
 
 class LoginController extends Controller {
