+++ lib/ECommerce/Tables/Cookies.php
@@ -1,8 +1,9 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Tables;
 
+use ECommerce\Site;
 
 class Cookies extends Table {
