+++ lib/ECommerce/Tables/Users.php
@@ -1,8 +1,9 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Tables;
 
+use ECommerce\Site;
 
 class Users extends Table {
