+++ lib/ECommerce/Views/AddSubCatView.php
@@ -1,32 +1,27 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Views;
 
 
-class AddTopCatView extends View {
+class AddSubCatView extends View {
 
     public function __construct() {
-        $this->setTitle("Add Top Category");
+        $this->setTitle("Add Sub-Category");
 
         $this->addLink("./staff.php", "Staff");
         $this->addLink("./users.php", "Users");
         $this->addLink("./profile.php", "Profile");
-        $this->addLink("./add-sub-cat.php", "Add Sub");
+        $this->addLink("./add-top-cat.php", "Add Top");
         $this->addLink("./add-product.php", "Add Product");
         $this->addLink("post/logout.php", "Log Out");
     }
 
     public function present() {
-        return $this->AddTop();
-
-    }
-
-    public function AddTop() {
         return <<<HTML
-<form id="add-top-cat">
+<form id="add-sub-cat" name="add-sub-cat">
     <fieldset>
-        <legend>Top Category</legend>
+        <legend>Sub Category</legend>
         <p>
             <label for="name">Name</label><br>
             <input type="text" id="name" name="name" placeholder="Name">
@@ -36,17 +31,21 @@ class AddTopCatView extends View {
             <textarea id="description" name="description" placeholder="Description..."></textarea>
         </p>
         <p>
+            <label for="file">Select Image to upload</label>
+            <input type="file" id="file" name="file">
+        <p>
             <label for="visible">Visible</label>
             <select id="visible" name="visible">
-                <option value="true">True</option>
-                <option value="false">False</option>
+                <option value="1">True</option>
+                <option value="0">False</option>
             </select>
         </p>
        	<p>
-			<input type="submit" value="OK">
+			<input type="submit" value="Upload Image" name="submit">
 		</p>
     </fieldset>
 </form>
 HTML;
     }
+
 }
