+++ lib/ECommerce/Views/AddTopCatView.php
@@ -1,27 +1,32 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Views;
 
 
-class AddSubCatView extends View {
+class AddTopCatView extends View {
 
     public function __construct() {
-        $this->setTitle("Add Sub-Category");
+        $this->setTitle("Add Top Category");
 
         $this->addLink("./staff.php", "Staff");
         $this->addLink("./users.php", "Users");
         $this->addLink("./profile.php", "Profile");
-        $this->addLink("./add-top-cat.php", "Add Top");
+        $this->addLink("./add-sub-cat.php", "Add Sub");
         $this->addLink("./add-product.php", "Add Product");
         $this->addLink("post/logout.php", "Log Out");
     }
 
     public function present() {
+        return $this->AddTop();
+
+    }
+
+    public function AddTop() {
         return <<<HTML
-<form id="add-sub-cat" name="add-sub-cat">
+<form id="add-top-cat">
     <fieldset>
-        <legend>Sub Category</legend>
+        <legend>Top Category</legend>
         <p>
             <label for="name">Name</label><br>
             <input type="text" id="name" name="name" placeholder="Name">
@@ -31,21 +36,17 @@ class AddSubCatView extends View {
             <textarea id="description" name="description" placeholder="Description..."></textarea>
         </p>
         <p>
-            <label for="file">Select Image to upload</label>
-            <input type="file" id="file" name="file">
-        <p>
             <label for="visible">Visible</label>
             <select id="visible" name="visible">
-                <option value="1">True</option>
-                <option value="0">False</option>
+                <option value="true">True</option>
+                <option value="false">False</option>
             </select>
         </p>
        	<p>
-			<input type="submit" value="Upload Image" name="submit">
+			<input type="submit" value="OK">
 		</p>
     </fieldset>
 </form>
 HTML;
     }
-
 }
