+++ lib/ECommerce/Views/LoginView.php
@@ -1,8 +1,10 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Views;
 
+use ECommerce\Site;
+use ECommerce\Tables\Cookies;
 
 class LoginView extends View {
 
@@ -41,7 +43,7 @@ class LoginView extends View {
         <p><input type="submit" value="Log in"></p>
         <p><a href="">Lost Password</a></p>
 
-        <p><a href="./">BT Glass Home</a></p>
+        <p><a href="..">BT Glass Home</a></p>
         <p><a href="./register.php">Register</a></p>
     </fieldset>
