+++ lib/ECommerce/Views/StaffView.php
@@ -1,8 +1,14 @@
 <?php
 
 
-namespace ECommerce;
+namespace ECommerce\Views;
 
+
+use ECommerce\Site;
+use ECommerce\Tables\Products;
+use ECommerce\Tables\SubCategories;
+use ECommerce\Tables\TopCategories;
+use ECommerce\Tables\Users;
 
 class StaffView extends View {
 
@@ -19,7 +25,7 @@ class StaffView extends View {
         $this->subCategories = new SubCategories($this->site);
         $this->products = new Products($this->site);
 
-        $this->setTitle("Staff View");
+        $this->setTitle("Staff Views");
         $this->addLink("./users.php", "Users");
         $this->addLink("./profile.php", "Profile");
         $this->addLink("post/logout.php", "Log Out");
